Complete overhaul of the renderer.
In the previous versions, triangles were drawn on the screen one by one, in a sequential manner. The renderer would completely finish drawing the current triangle before moving on to the next one. This is intuitive and works well on a single threaded machine. However, this is not how modern GPUs work at all. The new architecture kind of imitates how GPUs process triangles.